+++ packages/stream_chat_flutter/lib/src/scroll_view/thread_scroll_view/stream_thread_list_tile.dart
@@ -1,6 +1,5 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:stream_chat_flutter/src/channel/stream_draft_message_preview_text.dart';
 import 'package:stream_chat_flutter/src/misc/timestamp.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
@@ -260,6 +259,7 @@ class ThreadLatestReply extends StatelessWidget {
                   StreamTimestamp(
                     date: latestReply.createdAt.toLocal(),
                     style: theme.threadLatestReplyTimestampStyle,
+                    formatter: theme.threadLatestReplyTimestampFormatter,
                   ),
                 ],
               ),